+++ src/app/api/admin/orders/route.ts
@@ -4,53 +4,6 @@ import { FlyGuyErrorResponse } from "@/utils/errorResponse";
 
 import { OrdersController } from "@/app/controllers/ordersController";
 
-import { pool } from "../../../../config/db";
-
-import { getPrices } from "@/utils/getPrices";
-import { sumOrderItems } from "@/utils/sumOrderItems";
-
-export interface OrdersDTO {
-  orders: DatabaseOrder[];
-}
-
-export interface OrderDTO {
-  items: OrderItemDTO[];
-  id: number;
-  order_date: string;
-  customer_name: string;
-  user_id: string;
-  order_total: string;
-}
-
-export interface OrderItemDTO {
-  product_name: string;
-  quantity: string;
-  item_total: string;
-}
-
-export interface DatabaseOrder {
-  id: number;
-  order_date: string;
-  customer_name: string;
-  user_id: string;
-  order_total: string;
-  product_name: string;
-  quantity: string;
-  item_total: string;
-}
-
-export interface CreateOrderDTO {
-  user_id: string;
-  customer_name: string;
-  order_total: number;
-  order_items: CreateOrderItemDTO[];
-}
-
-export interface CreateOrderItemDTO {
-  id: string;
-  quantity: number;
-}
-
 export async function GET(req: NextRequest) {
   try {
     const queryParams = req.nextUrl.searchParams;
@@ -61,7 +14,6 @@ export async function GET(req: NextRequest) {
     }
     return await controller.getOrders();
   } catch (e) {
-    console.error("it errored");
     return FlyGuyErrorResponse();
   }
 }
